Add files via upload
Fixed issue where there is no checking if a bomb was in the tile or not before placing it in the same tile.
Added flip functionality, user can't flip if already flipped.  if flip bomb, then we put bomb.  If no bomb, user can continue guessing.
Row and column offsets to find out what the neighbor values are for a tile when a user flips a tile